+++ src/surgescript/runtime/variable.h
@@ -42,14 +42,14 @@ surgescript_var_t* surgescript_var_destroy(surgescript_var_t* var);
 /* retrieve the value stored in a variable */
 bool surgescript_var_is_null(const surgescript_var_t* var);
 bool surgescript_var_get_bool(const surgescript_var_t* var);
-float surgescript_var_get_number(const surgescript_var_t* var);
+double surgescript_var_get_number(const surgescript_var_t* var);
 char* surgescript_var_get_string(const surgescript_var_t* var, const struct surgescript_objectmanager_t* manager); /* warning: allocates a new buffer; you have to ssfree() this. See also: surgescript_var_to_string() */
 unsigned surgescript_var_get_objecthandle(const surgescript_var_t* var);
 
 /* sets the value of a variable */
 surgescript_var_t* surgescript_var_set_null(surgescript_var_t* var);
 surgescript_var_t* surgescript_var_set_bool(surgescript_var_t* var, bool boolean);
-surgescript_var_t* surgescript_var_set_number(surgescript_var_t* var, float number);
+surgescript_var_t* surgescript_var_set_number(surgescript_var_t* var, double number);
 surgescript_var_t* surgescript_var_set_string(surgescript_var_t* var, const char* string);
 surgescript_var_t* surgescript_var_set_objecthandle(surgescript_var_t* var, unsigned handle);
